Refactor patching CoreDNS
