+++ test/e2e/remote_dns_name/main_test.go
@@ -24,6 +24,7 @@ import (
 	"istio.io/istio/pkg/test/framework/components/namespace"
 
 	"github.com/openshift-service-mesh/federation/test/e2e/common"
+	"github.com/openshift-service-mesh/federation/test/e2e/common/coredns"
 )
 
 func TestMain(m *testing.M) {
@@ -34,7 +35,7 @@ func TestMain(m *testing.M) {
 		Setup(common.RecreateControlPlaneNamespace).
 		Setup(common.CreateCACertsSecret).
 		Setup(common.DeployControlPlanes("k8s")).
-		Setup(common.PatchCoredns).
+		Setup(coredns.PatchHosts).
 		Setup(common.InstallOrUpgradeFederationControllers(common.InstallOptions{SetRemoteDNSName: true})).
 		Setup(namespace.Setup(&common.AppNs, namespace.Config{Prefix: "app", Inject: true})).
 		// a - client
